chore: extract unique id modifier as a separate feature
