+++ model/UniqueId/Listener/ItemUpdatedEventListener.php
@@ -20,50 +20,59 @@
 
 declare(strict_types=1);
 
-namespace oat\taoQtiItem\model\Translation\Form\Modifier;
+namespace oat\taoQtiItem\model\UniqueId\Listener;
 
 use oat\generis\model\data\Ontology;
 use oat\tao\model\featureFlag\FeatureFlagCheckerInterface;
-use oat\tao\model\form\Modifier\AbstractFormModifier;
 use oat\tao\model\TaoOntology;
-use oat\taoQtiItem\model\Translation\Service\QtiIdentifierRetriever;
-use tao_helpers_form_Form;
-use tao_helpers_Uri;
+use oat\taoItems\model\event\ItemUpdatedEvent;
+use oat\taoQtiItem\model\UniqueId\Service\QtiIdentifierRetriever;
+use Psr\Log\LoggerInterface;
 
-class TranslationFormModifier extends AbstractFormModifier
+class ItemUpdatedEventListener
 {
+    private FeatureFlagCheckerInterface $featureFlagChecker;
     private Ontology $ontology;
     private QtiIdentifierRetriever $qtiIdentifierRetriever;
-    private FeatureFlagCheckerInterface $featureFlagChecker;
+    private LoggerInterface $logger;
 
     public function __construct(
+        FeatureFlagCheckerInterface $featureFlagChecker,
         Ontology $ontology,
         QtiIdentifierRetriever $qtiIdentifierRetriever,
-        FeatureFlagCheckerInterface $featureFlagChecker
+        LoggerInterface $logger
     ) {
+        $this->featureFlagChecker = $featureFlagChecker;
         $this->ontology = $ontology;
         $this->qtiIdentifierRetriever = $qtiIdentifierRetriever;
-        $this->featureFlagChecker = $featureFlagChecker;
+        $this->logger = $logger;
     }
 
-    public function modify(tao_helpers_form_Form $form, array $options = []): void
+    public function populateUniqueId(ItemUpdatedEvent $event): void
     {
-        if (!$this->featureFlagChecker->isEnabled('FEATURE_FLAG_TRANSLATION_ENABLED')) {
+        if (!$this->featureFlagChecker->isEnabled('FEATURE_FLAG_UNIQUE_NUMERIC_QTI_IDENTIFIER')) {
             return;
         }
 
-        $encodedProperty = tao_helpers_Uri::encode(TaoOntology::PROPERTY_UNIQUE_IDENTIFIER);
-        $uniqueIdValue = $form->getValue($encodedProperty);
+        $uniqueIdProperty = $this->ontology->getProperty(TaoOntology::PROPERTY_UNIQUE_IDENTIFIER);
+        $item = $this->ontology->getResource($event->getItemUri());
 
-        if (!empty($uniqueIdValue)) {
+        if (!empty((string) $item->getOnePropertyValue($uniqueIdProperty))) {
+            $this->logger->info(
+                sprintf(
+                    'The property "%s" for the item "%s" has already been set.',
+                    $uniqueIdProperty->getUri(),
+                    $item->getUri()
+                )
+            );
+
             return;
         }
 
-        $instance = $this->ontology->getResource($form->getValue('uri'));
-        $identifier = $this->qtiIdentifierRetriever->retrieve($instance);
+        $identifier = $this->qtiIdentifierRetriever->retrieve($item);
 
         if ($identifier) {
-            $form->setValue($encodedProperty, $identifier);
+            $item->setPropertyValue($uniqueIdProperty, $identifier);
         }
     }
 }
