+++ test/unit/model/Translation/Listener/ItemUpdatedEventListenerTest.php
@@ -28,8 +28,8 @@ use oat\generis\model\data\Ontology;
 use oat\tao\model\featureFlag\FeatureFlagCheckerInterface;
 use oat\tao\model\TaoOntology;
 use oat\taoItems\model\event\ItemUpdatedEvent;
-use oat\taoQtiItem\model\Translation\Listener\ItemUpdatedEventListener;
-use oat\taoQtiItem\model\Translation\Service\QtiIdentifierRetriever;
+use oat\taoQtiItem\model\UniqueId\Listener\ItemUpdatedEventListener;
+use oat\taoQtiItem\model\UniqueId\Service\QtiIdentifierRetriever;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
